Minor error fixing
Added stubs for socket listener, changed signature of connection handler
thread routine

+++ Atroshenko/config.h
@@ -2,7 +2,7 @@
 #define SERVICETASK_CONFIG
 
 struct configuration {
-	short listen_port;
+	unsigned int listen_port;
 	char *target_path;
 	char *log_path;
 };
